Add support for WMS servers, LandscapeTexturer, distance blend

+++ Source/FileDownloader/Public/FileDownloader/Download.h
@@ -10,7 +10,7 @@ public:
 	static bool SynchronousFromURLExpecting(FString URL, FString File, int32 ExpectedSize);
 	static bool SynchronousFromURL(FString URL, FString File);
 
-	static void FromURLExpecting(FString URL, FString File, int32 ExpectedSize, TFunction<void(bool)> OnComplete);
+	static void FromURLExpecting(FString URL, FString File, int64 ExpectedSize, TFunction<void(bool)> OnComplete);
 	static void FromURL(FString URL, FString File, TFunction<void(bool)> OnComplete);
 
 	static void AddExpectedSize(FString URL, int32 ExpectedSize);
